requirements.txt
torch==2.8.0
numpy>=1.26
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
PyYAML>=6.0
tqdm>=4.66
